ai/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0

pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

SQLAlchemy==2.0.45
psycopg[binary]==3.2.3
pgvector==0.4.2

kafka-python==2.2.15

openai>=1.0.0
httpx>=0.27.0
tenacity>=8.2.3